+++ src/core/lombok/extern/jackson/Jacksonized.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 The Project Lombok Authors.
+ * Copyright (C) 2020-2025 The Project Lombok Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -28,12 +28,21 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
 import lombok.Builder;
+import lombok.experimental.Accessors;
 import lombok.experimental.SuperBuilder;
 
 /**
  * The {@code @Jacksonized} annotation is an add-on annotation for
- * {@code @}{@link Builder} and {@code @}{@link SuperBuilder}. It automatically
- * configures the generated builder class to be used by Jackson's
+ * {@code @}{@link Builder}, {@code @}{@link SuperBuilder}, and
+ * {@code @}{@link Accessors}.
+ * <p>
+ * For {@code @}{@link Accessors}{@code (fluent = true)} on a type, it automatically
+ * configures Jackson to use the generated setters and getters for
+ * (de-)serialization by inserting {@code @}{@link JsonProperty} annotations.
+ * (Note that {@code @Jacksonized} {@code @Accessors} on fields are not supported.) 
+ * <p>
+ * For {@code @}{@link Builder} and {@code @}{@link SuperBuilder}, it
+ * automatically configures the generated builder class to be used by Jackson's
  * deserialization. It only has an effect if present at a context where there is
  * also a {@code @Builder} or a {@code @SuperBuilder}; a warning is emitted
  * otherwise.
@@ -41,8 +50,8 @@ import lombok.experimental.SuperBuilder;
  * In particular, the annotation does the following:
  * <ul>
  * <li>Configure Jackson to use the builder for deserialization using
- * {@code @JsonDeserialize(builder=Foobar.FoobarBuilder[Impl].class)}
- * on the class (where <em>Foobar</em> is the name of the annotated class).</li>
+ * {@code @JsonDeserialize(builder=Foobar.FoobarBuilder[Impl].class)} on the
+ * class (where <em>Foobar</em> is the name of the annotated class).</li>
  * <li>Copy Jackson-related configuration annotations (like
  * {@code @JsonIgnoreProperties}) from the class to the builder class. This is
  * necessary so that Jackson recognizes them when using the builder.</li>
@@ -54,9 +63,9 @@ import lombok.experimental.SuperBuilder;
  * <li>For {@code @SuperBuilder}, make the builder implementation class
  * package-private.</li>
  * </ul>
- * This annotation does <em>not</em> change the behavior of the generated builder.
- * A {@code @Jacksonized} {@code @SuperBuilder} remains fully compatible to
- * regular {@code @SuperBuilder}s.
+ * This annotation does <em>not</em> change the behavior of the generated
+ * builder. A {@code @Jacksonized} {@code @SuperBuilder} remains fully
+ * compatible to regular {@code @SuperBuilder}s.
  */
 @Target({TYPE, METHOD, CONSTRUCTOR})
 @Retention(SOURCE)